Refactor out endpoints related to static assets, including Auspice and Gatsby
The ultimate goal is to tidy up app.js to be mostly a thin file wiring
up routes (i.e. /foo/:bar) to endpoints, so that growing the number of
routes and the internal complexity of endpoints doesn't make it hard to
see the overall structure of the app.
